// client/app/src/generated/dto/social.ts
export interface ChatMessage {
    /**
     * @generated from protobuf field: int32 id = 5
     */
    id: number;
    /**
     * @generated from protobuf field: int32 sender_id = 1
     */
    senderId: number;
    /**
     * @generated from protobuf field: int32 receiver_id = 2
     */
    receiverId: number;
    /**
     * @generated from protobuf field: string content = 3
     */
    content: string;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp created_at = 4
     */
    createdAt?: Timestamp;
}
// ...
class ChatMessage$Type extends MessageType<ChatMessage> {
    constructor() {
        super("ChatMessage", [
            { no: 5, name: "id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 1, name: "sender_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 2, name: "receiver_id", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "content", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "created_at", kind: "message", T: () => Timestamp }
        ]);
    }
    create(value?: PartialMessage<ChatMessage>): ChatMessage {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.id = 0;
        message.senderId = 0;
        message.receiverId = 0;
        message.content = "";
        if (value !== undefined)
            reflectionMergePartial<ChatMessage>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: ChatMessage): ChatMessage {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* int32 id */ 5:
                    message.id = reader.int32();
                    break;
                case /* int32 sender_id */ 1:
                    message.senderId = reader.int32();
                    break;
                case /* int32 receiver_id */ 2:
                    message.receiverId = reader.int32();
                    break;
                case /* string content */ 3:
                    message.content = reader.string();
                    break;
                case /* google.protobuf.Timestamp created_at */ 4:
                    message.createdAt = Timestamp.internalBinaryRead(reader, reader.uint32(), options, message.createdAt);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: ChatMessage, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* int32 sender_id = 1; */
        if (message.senderId !== 0)
            writer.tag(1, WireType.Varint).int32(message.senderId);
        /* int32 receiver_id = 2; */
        if (message.receiverId !== 0)
            writer.tag(2, WireType.Varint).int32(message.receiverId);
        /* string content = 3; */
        if (message.content !== "")
            writer.tag(3, WireType.LengthDelimited).string(message.content);
        /* google.protobuf.Timestamp created_at = 4; */
        if (message.createdAt)
            Timestamp.internalBinaryWrite(message.createdAt, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        /* int32 id = 5; */
        if (message.id !== 0)
            writer.tag(5, WireType.Varint).int32(message.id);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message ChatMessage
 */
export const ChatMessage = new ChatMessage$Type();